Updated state and props for sales details form. User data entry was not saving.
